backend/requirements.txt
# Core backend framework
fastapi==0.110.0
uvicorn==0.29.0

# Data validation & schemas
# pydantic==2.6.4

# File uploads
python-multipart==0.0.9

# Different file formats
pypdf
python-docx
pandas

# Async HTTP (for agent-to-agent calls & LLM APIs)
httpx==0.27.0

# Google Gemini SDK
google-generativeai==0.4.1

# Environment variable management
python-dotenv==1.0.1

# Optional: logging & utilities
rich==13.7.1
